refactor(build-optimizer): move containers and components in seperated folders
